-all routing jalan -authentication selesai

+++ controller/usersController.js
@@ -1,3 +1,8 @@
+const bcrypt 	 = require('bcrypt');
+const saltRounds = 10;
+const jwt 	     = require('jsonwebtoken');
+require('dotenv').config()
+
 const User = require('../models').User
 
 const findAllUser = (req, res) => {
@@ -32,30 +37,40 @@ const createUser = (req, res) => {
 }
 
 const deleteUser = (req, res) => {
-	User.destroy({
-		where: {
-			id: req.params.id
-		}
-	})
-	.then(() => {
-		res.send('user already gone')
-	})
-	.catch(err => {
-		res.send(err)
-	})
+	if(req.headers.UserId == req.params.id){
+		res.send('Permission Denied...')
+	}else{
+		User.destroy({
+			where: {
+				id: req.params.id
+			}
+		})
+		.then(() => {
+			res.send('user already gone')
+		})
+		.catch(err => {
+			res.send(err)
+		})
+	}
 }
 
 const updateUser = (req, res) => {
-	User.update({
-		name : req.body.name,
-		age : req.body.age,
-		email : req.body.email,
-		isAdmin : req.body.isAdmin
-	},{where: {id: req.params.id}}).then((user, update) => {
-		res.send(update)
-	}).catch(err => {
-		res.send(err)
-	})
+	bcrypt.hash(req.body.password, saltRounds).then(function(hash) {
+      	User.update({
+			name : req.body.name,
+			age : req.body.age,
+			email : req.body.email,
+			isAdmin : req.body.isAdmin,
+			password : hash
+	   	},{
+	   		where: {
+	   			id: req.params.id}})
+      	.then((user, update) => {
+			res.send(update)
+		}).catch(err => {
+			res.send(err)
+		})
+    })
 }
 
 const signupUser = (req, res) => {
@@ -67,6 +82,7 @@ const signupUser = (req, res) => {
 		password : req.body.password
 	})
 	.then((user, create) => {
+		console.log(create);
 		res.send(user)
 	})
 	.catch(err => {
@@ -81,11 +97,18 @@ const signinUser = (req, res) => {
 		}
 	}).then(user => {
 		if(user){
-			if(user.password === req.body.password){
-				res.send('Yeay Login')
-			}else{
-				res.send('Email or Password is invalid')
-			}
+			bcrypt.compare(req.body.password, user.password).then(function(result) {
+   				if(result){
+   					jwt.sign({UserId : user.id, name : user.name, email : user.email, isAdmin : user.isAdmin, isLogin : true}, process.env.Salt_word, (err, token)=> {
+   						req.headers.token = token
+   						res.send(token)
+   					})
+				}else{
+					res.send('Email or Password is invalid')
+				}
+			}).catch(err => {
+				res.send(err)
+			})
 		}else{
 			res.send('Email or Password is invalid')
 		}
@@ -101,5 +124,6 @@ module.exports = {
 	createUser,
 	deleteUser,
 	updateUser,
-	signupUser
+	signupUser,
+	signinUser
 }
